move getShowAttrDflt function to a separate file
