+++ src/plots/cartesian/tick_label_defaults.js
@@ -3,6 +3,7 @@
 var Lib = require('../../lib');
 var contrast = require('../../components/color').contrast;
 var layoutAttributes = require('./layout_attributes');
+var getShowAttrDflt = require('./show_dflt');
 var handleArrayContainerDefaults = require('../array_container_defaults');
 
 module.exports = function handleTickLabelDefaults(containerIn, containerOut, coerce, axType, options, config) {
@@ -66,34 +67,6 @@ function handleOtherDefaults(containerIn, containerOut, coerce, axType, options)
                 coerce('separatethousands');
             }
         }
-    }
-}
-
-/*
- * Attributes 'showexponent', 'showtickprefix' and 'showticksuffix'
- * share values.
- *
- * If only 1 attribute is set,
- * the remaining attributes inherit that value.
- *
- * If 2 attributes are set to the same value,
- * the remaining attribute inherits that value.
- *
- * If 2 attributes are set to different values,
- * the remaining is set to its dflt value.
- *
- */
-function getShowAttrDflt(containerIn) {
-    var showAttrsAll = ['showexponent', 'showtickprefix', 'showticksuffix'];
-    var showAttrs = showAttrsAll.filter(function(a) {
-        return containerIn[a] !== undefined;
-    });
-    var sameVal = function(a) {
-        return containerIn[a] === containerIn[showAttrs[0]];
-    };
-
-    if(showAttrs.every(sameVal) || showAttrs.length === 1) {
-        return containerIn[showAttrs[0]];
     }
 }
